moved convertPrecision helper into bignumber helper file
