+++ src/lib/bignumber/bignumber.ts
@@ -15,3 +15,14 @@ export const positiveOrZero = (n: BigNumber.Value | null | undefined): BN => {
   const value = bn(n || 0)
   return value.isPositive() ? value : bn(0)
 }
+
+export const convertPrecision = (
+  value: BigNumber.Value,
+  inputPrecision: number,
+  outputPrecision: number,
+): BigNumber => {
+  return bnOrZero(value)
+    .dividedBy(bn(10).exponentiatedBy(inputPrecision))
+    .times(bn(10).exponentiatedBy(outputPrecision))
+    .integerValue()
+}
